+++ backend/models/employee.py
@@ -12,17 +12,22 @@ class Employee(BasicModel):
         employees = self.collection.find(sort=[('name', 1)])
         employees_list = []
         for employee in employees:
-            if 'email' in employee and employee['email']:
-                email = employee['email']
-            else:
-                email = None
+
             # 퇴사하지 않은 직원만 포함하기 위해서
             if 'regular' not in employee:
                 employee['regular'] = '상근'
+                
             if employee['regular'] != '퇴사':
-                if not 'mode' in employee:
+                if 'email' in employee and employee['email']:
+                    email = employee['email']
+                else:
+                    email = None            
+                if 'mode' not in employee:
                     employee['mode'] = '내근'
-                employee_info = {'name': employee['name'], 'employeeId': employee['employeeId'], 'email': email, 'regular': employee['regular'], 'mode': employee['mode']}
+                if 'attendMode' not in employee:
+                    employee['attendMode'] = 'O'
+
+                employee_info = {'name': employee['name'], 'employeeId': employee['employeeId'], 'email': email, 'regular': employee['regular'], 'mode': employee['mode'], 'attendMode': employee['attendMode']}
                 if 'beginDate' not in employee and 'endDate' not in employee:
                     employees_list.append(employee_info)
                 elif 'beginDate' in employee and date and date >= employee['beginDate'] and 'endDate' not in employee: