+++ client/src/utils/AuthUtil.jsx
@@ -9,10 +9,10 @@ export const requestAuth = async (mode, method, value, dispatch, navigate, setEr
             const res = await axios.post(url, value)
             if (mode === 'login') {
                 dispatch(loginUser(res.data)) 
-                if (value.location) {
+                if (res.data.where.attend || res.data.where.isMobile === 'O') {
                     navigate('/', {state : {location: value.location, where: res.data.where}})
                 } else {
-                    navigate('/')
+                    navigate('/attend')
                 }
             } else if (mode === 'register') {
                 navigate('/check-email')